pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "google-adk-a2a-agents"
version = "0.1.0"
description = "Google ADK multi-agent system with A2A protocol"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.3.0",
    "google-cloud-aiplatform>=1.60.0",
    "google-cloud-bigquery>=3.20.0",
    "google-auth>=2.30.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

